+++ lib/tornado/platform/common.py
@@ -1,5 +1,5 @@
 """Lowest-common-denominator implementations of platform functionality."""
-from __future__ import absolute_import, division, print_function, with_statement
+from __future__ import absolute_import, division, print_function
 
 import errno
 import socket
@@ -7,6 +7,7 @@ import time
 
 from tornado.platform import interface
 from tornado.util import errno_from_exception
+
 
 def try_close(f):
     # Avoid issue #875 (race condition when using the file in another